Backend/requirements.txt
fastapi>=0.95.2
uvicorn>=0.22.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0
streamlit>=1.28.0
langchain-huggingface>=0.0.4
langchain-chroma>=0.0.4
langchain-community>=0.0.18
chromadb>=0.4.15
sentence-transformers>=2.2.2
pymupdf>=1.22.5
unstructured>=0.10.8
python-multipart>=0.0.6
pypdf>=3.15.0
langchain-core>=0.1.0
huggingface-hub>=0.19.0
sqlalchemy>=2.0.0